refactor(artifact-provenance): extract execution evidence
