+++ scripts/ci/module-impact.json
@@ -145,6 +145,8 @@
     },
     "artifact_provenance": {
       "ownerPaths": [
+        "src/main/artifacts/provenance-canonical.ts",
+        "src/main/artifacts/provenance-execution-evidence.ts",
         "src/main/artifacts/provenance-message-snapshot.ts",
         "src/main/artifacts/provenance-repository.ts",
         "src/main/artifacts/provenance-version-writer.ts"
